Conflicts:
	src/pouch.js

+++ src/adapters/pouch.http.js
@@ -198,6 +198,10 @@ var HttpPouch = function(opts, callback) {
   };
 
   api.request = function(options, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('request', arguments);
+      return;
+    }
     options.auth = host.auth;
     options.url = genDBUrl(host, options.url);
     ajax(options, callback);
@@ -206,6 +210,10 @@ var HttpPouch = function(opts, callback) {
   // Sends a POST request to the host calling the couchdb _compact function
   //    version: The version of CouchDB it is running
   api.compact = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('compact', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       url: genDBUrl(host, '_compact'),
@@ -217,6 +225,10 @@ var HttpPouch = function(opts, callback) {
   //    couchdb: A welcome string
   //    version: The version of CouchDB it is running
   api.info = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('info', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       method:'GET',
@@ -228,6 +240,10 @@ var HttpPouch = function(opts, callback) {
   // The id could be solely the _id in the database, or it may be a
   // _design/ID or _local/ID path
   api.get = function(id, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('get', arguments);
+      return;
+    }
     // If no options were given, set the callback to the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -304,8 +320,8 @@ var HttpPouch = function(opts, callback) {
     var parts = id.split('/');
     if ((parts.length > 1 && parts[0] !== '_design' && parts[0] !== '_local') ||
         (parts.length > 2 && parts[0] === '_design' && parts[0] !== '_local')) {
-      // Nothing is expected back from the server
-      options.json = false;
+      // Binary is expected back from the server
+      options.binary = true;
     }
 
     // Get the document
@@ -322,6 +338,10 @@ var HttpPouch = function(opts, callback) {
 
   // Delete the document given by doc from the database given by host.
   api.remove = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('remove', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -338,6 +358,10 @@ var HttpPouch = function(opts, callback) {
 
   // Remove the attachment given by the id and rev
   api.removeAttachment = function idb_removeAttachment(id, rev, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('removeAttachment', arguments);
+      return;
+    }
     ajax({
       auth: host.auth,
       method: 'DELETE',
@@ -345,23 +369,47 @@ var HttpPouch = function(opts, callback) {
     }, callback);
   };
 
-  // Add the attachment given by doc and the content type given by type
+  // Add the attachment given by blob and its contentType property
   // to the document with the given id, the revision given by rev, and
   // add it to the database given by host.
-  api.putAttachment = function(id, rev, doc, type, callback) {
+  api.putAttachment = function(id, rev, blob, type, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('putAttachment', arguments);
+      return;
+    }
+    if (typeof type === 'function') {
+      callback = type;
+      type = blob;
+      blob = rev;
+      rev = null;
+    }
+    if (typeof type === 'undefined') {
+      type = blob;
+      blob = rev;
+      rev = null;
+    }
+    var url = genDBUrl(host, id);
+    if (rev) {
+      url += '?rev=' + rev;
+    }
     // Add the attachment
     ajax({
       auth: host.auth,
       method:'PUT',
-      url: genDBUrl(host, id) + '?rev=' + rev,
+      url: url,
       headers: {'Content-Type': type},
-      body: doc
+      processData: false,
+      body: blob
     }, callback);
   };
 
   // Add the document given by doc (in JSON string format) to the database
   // given by host. This fails if the doc has no _id field.
   api.put = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('put', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -401,6 +449,10 @@ var HttpPouch = function(opts, callback) {
   // given by host. This does not assume that doc is a new document (i.e. does not
   // have a _id or a _rev field.
   api.post = function(doc, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('post', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -427,6 +479,10 @@ var HttpPouch = function(opts, callback) {
   // Update/create multiple documents given by req in the database
   // given by host.
   api.bulkDocs = function(req, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('bulkDocs', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -458,6 +514,10 @@ var HttpPouch = function(opts, callback) {
   // by host and ordered by increasing id.
   api.allDocs = function(opts, callback) {
     // If no options were given, set the callback to be the second parameter
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('allDocs', arguments);
+      return;
+    }
     if (typeof opts === 'function') {
       callback = opts;
       opts = {};
@@ -528,6 +588,10 @@ var HttpPouch = function(opts, callback) {
   // TODO According to the README, there should be two other methods here,
   // api.changes.addListener and api.changes.removeListener.
   api.changes = function(opts) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('changes', arguments);
+      return;
+    }
 
     if (Pouch.DEBUG) {
       console.log(db_url + ': Start Changes Feed: continuous=' + opts.continuous);
@@ -686,6 +750,10 @@ var HttpPouch = function(opts, callback) {
   // those that do NOT correspond to revisions stored in the database.
   // See http://wiki.apache.org/couchdb/HttpPostRevsDiff
   api.revsDiff = function(req, opts, callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('revsDiff', arguments);
+      return;
+    }
     // If no options were given, set the callback to be the second parameter
     if (typeof opts === 'function') {
       callback = opts;
@@ -699,11 +767,15 @@ var HttpPouch = function(opts, callback) {
       url: genDBUrl(host, '_revs_diff'),
       body: req
     }, function(err, res) {
-      call(callback, null, res);
+      call(callback, err, res);
     });
   };
 
   api.close = function(callback) {
+    if (!api.taskqueue.ready()) {
+      api.taskqueue.addJob('close', arguments);
+      return;
+    }
     call(callback, null);
   };
